prog.py: Fix the tcsetattr() ispeed argument
According to man termios(3):

cfsetispeed()  sets the input baud rate stored in the termios structure
to speed, which must be specified as one of the Bnnn constants listed
above for cfsetospeed().  If the input baud rate is set to zero, the
input baud rate will be equal to the output baud rate.

Apparently this isn't the case for Python's tcsetattr() interface...