bootstrap: Fix a sign-extension problem in the ihex address
Also increase the stdio input buffer from 32 to 64 bytes.  Since
programming is coming from automation in chunks up to 43 characters in
length now, I'm worried about it coming too fast for the REPL to keep up
with.  As long as rx_buf[] is larger than 43, it shouldn't be a problem.