bootstrap: Implement SPI using software
The P89V51RD2's hardware SPI implementation seems to have a bug that the
last bit sent isn't always held at its value for the final clock edge.
For example, the following depicts how transmission of 0x55 looks on an
oscilloscope:

        _   _   _   _   _   _   _   _
SCK  __| |_| |_| |_| |_| |_| |_| |_| |__
          ___     ___     ___     _
MOSI ____|   |___|   |___|   |___| |____
       0   1   0   1   0   1   0   ?

It's as though on the last rising edge of SCK, the shift register wraps
around and the final bit transitions to the value of the first bit;
e.g., 0x55 is intermittently received as 0x54, 0xaa as 0xab, 0x01 as
0x00, and 0x80 as 0x81, but 0x5a, 0xa5, 0x00 and 0x81 will all be sent
correctly because their LSB and MSB are the same.

I don't know whether this is a defect in the device or the fault of my
software not configuring something properly, but in the meantime a
simple software implementation of the SPI works just fine (although at a
significant impact to performance - from 76.8KHz to roughly 15.625KHz).